feat(healing): show each healer's active buffs in the per-round status panel
